// src/types.ts
export interface StockLot {
  lotNo: string;
  malzemeKodu: string;
  firma: string;
  kalanMiktar: number;
  ilkGirisMiktari: number;
  girisTarihi: string;
  depoLokasyonu?: string;
}